I want to commit this even though it's seriously broken.
Things work, *sometimes*. The client gets a lobby list but using a stream (TCP) connection
was a bad idea because sometimes I read more than a single message off the buffer. The server
and clients are exchanging small, distinct messages. Even though I want
this setup stuff to be reliable, it's not right to use a single TCP stream for all of it. I have
to either parse the stream for different messages or somehow add my own reliability layer to UDP.